Cellseg

+++ scripts/pre_process.py
@@ -1,12 +1,13 @@
-import numpy as np
-import tifffile as tiff
-import pandas as pd
 import os
-from orig_tiff import OrigTiff
-import shutil
 import sys
+import numpy as np
+import pandas as pd
+import tifffile as tiff
+import shutil
+from orig_tiff import OrigTiff
 sys.path.append('CellSeg')
 from CellSeg import main
+
 
 def cell_seg_prep(img:tiff, output_dir, numx:int,numy:int,channel_names):
 	'''Prepares image for CellSeg including tiling. Returns output_dir path that includes tiles and output subfolders.
@@ -16,7 +17,10 @@ def cell_seg_prep(img:tiff, output_dir, numx:int,numy:int,channel_names):
 	output_dir -- str -- path to directory one up of desired tiles and CellSeg output directories.
 	numx -- int -- number of tiles to divide image into on x-axis
 	numy -- number of tiles to divide image into on y-axis (for example, if numx=4 and numy=2 the image will be cropped into tiles by a 4x2 grid)
-	channel_names -- str -- path to text file that includes each image channel name as a line in order'''
+	channel_names -- str -- path to text file that includes each image channel name as a line in order
+	
+	returns:
+	target -- str -- path to project directory'''
 
 
 	if not os.path.exists(output_dir):
@@ -28,7 +32,8 @@ def cell_seg_prep(img:tiff, output_dir, numx:int,numy:int,channel_names):
 
 	tif = OrigTiff(img,numx,numy)
 	tif.crop_tif("%s/tiles" % output_dir)
-	shutil.copyfile(channel_names, "%s/channelNames.txt" % output_dir)
+	if not os.path.exists("%s/channelNames.txt" % output_dir):
+		shutil.copyfile(channel_names, "%s/channelNames.txt" % output_dir)
 	target = output_dir
 	return target
 
@@ -45,24 +50,27 @@ def run_cell_seg(img:tiff, output_dir, numx:int,numy:int,channel_names):
 	target = cell_seg_prep(img,output_dir,numx,numy,channel_names)
 	main.main(target)
 
-def merge_coords(coord_dir, file_names, xbounds:float, ybounds:float, x_name:str, y_name:str, output_dir: str):
+def merge_coords(coord_dir, tile_names, xbounds:float, ybounds:float, x_name:str, y_name:str, output_dir: str):
 	'''Converts CellSeg-segmetned tile cell center coordinates to coordinates in original image. Returns path to csv that contains cell coordiantes for the entire image.
 
 	inputs:
 	coord_dir -- str -- path to directory that contains CellSeg coordinate output files.
-	files_names -- list of str -- list of coordinate file names, using the OrigTiff tile naming convention
+	tile_names -- list of str -- list of image tile names, using the OrigTiff tile naming convention
 	xbounds -- list of float-- list of pixel x values for tile boundaries in terms of the entire image
 	ybounds -- list of float -- list of pixel y values for tile boundaries in terms of the entire image
 	x_name -- name of column in the coordinate file that contains cell center x coordinates in terms of the individual tile
 	y_name -- name of column in the coordinate file that contains cell center y coordinates in terms of the individual tile
-	output_dir -- str -- path to directory to output full coordinate file'''
+	output_dir -- str -- path to directory to output full coordinate file'
+	
+	returns:
+	path -- str -- path to file with all cell coordinates'''
 	isExist = os.path.exists(output_dir)
 	if not isExist:
 		os.makedirs(output_dir)
 	X = []
 	Y = []
-	for name in files:
-		data = pd.read_csv("%s/%s__statistics_growth2_uncomp.csv" % (coord_dir, file_name), usecols=[x_name,y_name])
+	for name in tile_names:
+		data = pd.read_csv("%s/%s__statistics_growth2_uncomp.csv" % (coord_dir, name), usecols=[x_name,y_name])
 		x_offset = xbounds[int(name[3]-1)]
 		y_offset = ybounds[int(name[2]-1)]
 		x = np.array(data[x_name]) + x_offset
@@ -77,7 +85,11 @@ def merge_coords(coord_dir, file_names, xbounds:float, ybounds:float, x_name:str
 	return path
 
 
-def crop_cell(img: np.array, x: int or float, y: int or float, box_size: int, idx: int, output_dir: str = "None", add_channels=False, num_channels:int) -> np.array:
+
+def run_cell_seg(img:tiff, output_dir, numx:int,numy:int,channel_names):
+	'''Prepares image for CellSeg and runs CellSeg using the submodule.
+
+def crop_cell(img: np.array, x: int or float, y: int or float, box_size: int, idx: int, num_channels:int, output_dir: str = "None",  add_channels=False) -> np.array:
 	'''Takes in multichannel img array and crops cell at specified center coordinates as an  individual image with a desired box size. Returns cropped image array. 
 	Can add specified number of empty (zero-filled) image channels to be compatible with pre-trained multi-channel vision transformer models.
 
@@ -89,7 +101,11 @@ def crop_cell(img: np.array, x: int or float, y: int or float, box_size: int, id
 	idx -- int -- cell index for naming output image
 	output_dir -- str -- output directory path (optional-if set to none the image array will be returned but the .tif image will not be saved.
 	add_channels -- bool -- If true, adds num_channel number of zero-filled channels to image. 
-	num_channels -- int -- Number of zero-filled channels to add to image. '''
+	num_channels -- int -- Number of zero-filled channels to add to image. 
+	
+	returns: 
+	crop -- np array -- cropped image array for single cell'''
+
 	#Get rectangle boundaries
 	x = np.round(x)
 	y = np.round(y)
@@ -105,10 +121,10 @@ def crop_cell(img: np.array, x: int or float, y: int or float, box_size: int, id
 		zeros = np.zeros((num_channels, box_size, box_size))
 		crop = np.concatenate((crop, zeros))
 	if output_dir != "None":
-		tiff.imwrite("%s/cell_%i.tif" % (output_dir, idx), crop, imagej=True)
+		tiff.imwrite("%s/cell_%i.tif" % (output_dir, idx), crop.astype('float32'), imagej=True)
 	return crop
 
-def img_to_cells(img: tiff, segs, box_size: int, output_dir: str, add_channels=False, num_channels:int):
+def img_to_cells(img: tiff, segs, box_size: int, output_dir: str, num_channels:int,add_channels=False):
 	'''Takes in pre-segmented multichannel and saves each cell as an individual image in tiff format of a specified box size.
 	Can add specified number of empty (zero-filled) image channels to be compatible with pre-trained multi-channel vision transformer models.
 
@@ -123,9 +139,9 @@ def img_to_cells(img: tiff, segs, box_size: int, output_dir: str, add_channels=F
 	segs = segs["coords"]
 	img = tiff.imread(img)
 	for i in range(segs.shape[1]):
-		crop_cell(img, segs[0,i], segs[1,i], box_size, i, output_dir, add_channels, num_channels)
+		crop_cell(img, segs[0,i], segs[1,i], box_size, i, num_channels, output_dir, add_channels)
 
-def prep_dino(img:tiff, coord_dir, numx:int, numy:int, nuclei_channel:int, boxsize:int, output_dir: str, add_channels=False, num_channels:int):
+def prep_dino(img:tiff, coord_dir, numx:int, numy:int, nuclei_channel:int, boxsize:int, output_dir: str, num_channels:int, add_channels=False):
 	'''Crops tiff into single cell images for use in scDINO.
 	Can add specified number of empty (zero-filled) image channels to be compatible with pre-trained scDINO 5 channel model.
 
@@ -140,11 +156,14 @@ def prep_dino(img:tiff, coord_dir, numx:int, numy:int, nuclei_channel:int, boxsi
 	output_dir -- str -- path to output directory for cell crops
 	add_channels -- bool -- If true, adds num_channel number of zero-filled channels to image. 
 	num_channels -- int -- Number of zero-filled channels to add to image.'''
+	
+	isExist = os.path.exists(output_dir)
+	if not isExist:
+		os.makedirs(output_dir)
 
-
-	tif = OrigTiff(img,numx,numy, nuclei_channel)
+	tif = OrigTiff(img,numx,numy)
 	xbounds, ybounds = tif.get_bounds()
 	files = tif.get_tile_names()
 	coords = merge_coords(coord_dir, files,  xbounds, ybounds, "Absolute X", "Absolute Y", output_dir)
-	img_to_cells(img, coords, boxsize, output_dir, add_channels, num_channels)
+	img_to_cells(img, coords, boxsize, output_dir, num_channels, add_channels)
